feat(socket): emit order to driver list

+++ lib/features/shippers/home/models/order_model.dart
@@ -135,8 +135,8 @@ class Order {
         ')';
   }
 
-  double getToTalPrice() {
-    double result = 0;
+  int getToTalPrice() {
+    int result = 0;
     for (var x in orderItems) result += x.totalPrice;
     return result;
   }
